jenkins

+++ vars/nodejsEKS.groovy
@@ -58,12 +58,28 @@ def call(Map configMap){
 
             stage('Deploy'){
             steps{
-                sh """
-                    aws eks update-kubeconfig --region ${region} --name ${project}-dev
-                    cd helm
-                    sed -i 's/IMAGE_VERSION/${appVersion}/g' values.yaml
-                    helm install ${component} -n ${project} .
-                """
+                 script{
+                    releaseExists = sh(script: "helm list -A --short | grep -w ${component} || true", returnStdout: true).trim()
+                        if(releaseExists.isEmpty()){
+                            echo "${component} not installed yet, first time installation"
+                            sh """
+                                aws eks update-kubeconfig --region ${region} --name ${project}-dev
+                                cd helm
+                                sed -i 's/IMAGE_VERSION/${appVersion}/g' values.yaml
+                                helm install ${component} -n ${project} .
+                            """
+                        }
+                        else{
+                            echo "${component} exists, running upgrade"
+                            sh"""
+                                aws eks update-kubeconfig --region ${region} --name ${project}-dev
+                                cd helm
+                                sed -i 's/IMAGE_VERSION/${appVersion}/g' values.yaml
+                                helm upgrade ${component} -n ${project} .
+                                """          
+                        }
+                    }
+                
                 }  
             }
             /* stage('Nexus Artifact Upload'){
